dom/html/test/file_iframe_sandbox_c_if6.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 341604</title>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<script type="text/javascript">
  function ok(result, desc) {
    window.parent.ok_wrapper(result, desc);
    window.parent.postMessage({ok: result, desc: desc}, "*");
  }
  
  function doStuff() {
    ok(true, "a document sandboxed with allow-same-origin and allow-scripts should be same origin with its parent and able to run scripts " +
       "regardless of what kind of whitespace was used in its sandbox attribute");
  }
</script>
<body onLoad="doStuff()">
  I am sandboxed but with "allow-same-origin" and "allow-scripts"
</body>
</html>
